chore : view 코드에 java docs 스타일 주석 추가 및 패키지 궂구조 변경
