+++ src/view/javafx/JavaFxSelectYutResultScreen.java
